Proposal: plugin mechanism for sharing image cache between machines
As has been discussed extensively in #20316, Docker now no longer
restores parent image chains when pulling from a registry, and people
who relied on this to cache builds across multiple machines have had
their build times blow out.

In this PR, I propose a plugin mechanism for the daemon to request
cached images from a plugin. I believe a plugin is appropriate here
because there are several different ways such sharing might work
depending on the lifecycle of the machines being used.

I'm PR'ing this to get some feedback from the maintainers & community on
this idea. If people agree that this is the way forward, I'd like to
start working on this. It doesn't appear to me that the implementation
of the plugin surface would be especially difficult or disruptive - but
I want your thoughts on this :)

Signed-off-by: KJ Tsanaktsidis <kjtsanaktsidis@gmail.com>